+++ CHANGELOG.md
@@ -1,1 +1,18 @@
 # Changelog
+
+# [2.0.0](https://github.com/NullVoxPopuli/ember-deep-tracked/compare/v1.3.14...v2.0.0) (2022-08-21)
+
+
+### chore
+
+* change ember support ([0a570f8](https://github.com/NullVoxPopuli/ember-deep-tracked/commit/0a570f80820f8c350ed79ba91f1d80fb8299c3f5))
+* **support:** changes to minimum supported peers ([5916eca](https://github.com/NullVoxPopuli/ember-deep-tracked/commit/5916eca51c99c09ac9882cefead0ecb6f787847a))
+
+
+### BREAKING CHANGES
+
+* minimum supporetd ember-source is now 3.25
+* **support:** typescript and ember-auto-import changes
+ - ember-auto-import v2+ is now required for consuming apps
+ - typescript < 4.5 is no longer supported. Currently supported
+   typescript is 4.5, 4.6, 4.7